item.jsの追加（前のIssueの修正）
